Setup django before server start to access db from socketio
