+++ project/__init__.py
@@ -1,0 +1,5 @@
+import os
+from django import setup
+
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'project.settings')
+setup()
