+++ server.js
@@ -1,6 +1,6 @@
 const express = require('express')
 const app = express()
-app.use(require('cors')())
+// app.use(require('cors')())
 const server = require('http').Server(app)
 const proxy = require('express-http-proxy')
 const PORT = process.env.PORT || 3000
